pyproject.toml
[tool.poetry]
name = "strava-dependencies"
version = "0.1.0"
description = ""
authors = ["Albert WSL <dewatrigant.albert@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.2.0"
requests = "^2.31.0"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
